+++ .gitignore
@@ -1,4 +1,3 @@
-
 /tests/output/
 /changelogs/.plugin-cache.yaml
